Create mediaPlayer.js
The goal of this scrip is to allow for an interactive interface for users to view signage as a button, as opposed to using half-wake mode.

This script makes use of the Digital Signage features of the Room and Board series of Webex Room Devices.
Its set up to take list of objects and dynamically produces the xml for the user interface. 

Each button produced has a URL which will update on button press, place the system into half wake mode, then kick it out of half wake mode at the end of the duration set by the integrator
This interface provided is how a user would interact with the media provided by an integrator.